chore(nlp): Adds a TotalParamsProxy objective.
